requirements.txt
# Core dependencies for Python 3.12 - September 2025
numpy==2.1.0
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2

# NLP and embeddings
transformers==4.45.0
sentence-transformers==3.1.0
nltk==3.9
spacy==3.8.0

# API clients
openai==1.45.0
anthropic==0.35.0
google-generativeai==0.8.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.66.5
pytest==8.3.3
jupyter==1.1.0

# Statistical analysis
pingouin==0.5.5
statsmodels==0.14.3
